Move methods to fov.cpp
