+++ CMakeLists.txt
@@ -61,6 +61,7 @@ set(CORE_SOURCES
     filesystem.cpp
     fish.cpp
     foobar_save.cpp
+    fov.cpp
     get_card_info.cpp
     get_random_item_id.cpp
     get_random_npc_id.cpp
