+++ animation.cpp
@@ -5,6 +5,7 @@
 #include "config.hpp"
 #include "draw.hpp"
 #include "elona.hpp"
+#include "fov.hpp"
 #include "item.hpp"
 #include "item_db.hpp"
 #include "map.hpp"
