+++ draw.cpp
@@ -3,6 +3,7 @@
 #include "character.hpp"
 #include "config.hpp"
 #include "elona.hpp"
+#include "fov.hpp"
 #include "map.hpp"
 #include "variables.hpp"
